remove apache license and rename to tblib.py and fix rat-excludes
